Update to Gradle 6.6.1 and attempt to deal with many versioning issues.
Right now, when attempting to use Kotlin 1.4.0, ther eare issues when using `gradleApi()` due to the difference in Kotlin versions that can't be fixed by simply changing how the dependencies resolve due to Gradle's special `gradleApi()` dependency.
Some versions have not been brought to their latest, but this changeset brings forward most of the versions.
There are still some Dokka strangeness in its inability to link to external documentation, but a neweer Dokka version is out for Kotlin 1.4 and I don't think it is worth the effort to figure it out with the current Dokka in this changeset.